misc: rename example codes to example_test
