+++ kakaowork/ImageBlock_test.go
@@ -1,29 +1,11 @@
 package kakaowork_test
 
 import (
-	"encoding/json"
-	"fmt"
 	"github.com/JSYoo5B/convertago/kakaowork"
 	"github.com/go-playground/validator/v10"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
-
-func ExampleImageBlock() {
-	textBlock := kakaowork.ImageBlock{Url: "https://something.storage.host/upload/path/filename"}
-
-	jsonBytes, err := json.MarshalIndent(textBlock, "", "    ")
-	if err != nil {
-		fmt.Println("Error:", err)
-		return
-	}
-
-	fmt.Println(string(jsonBytes))
-	// Output: {
-	//     "type": "image_link",
-	//     "url": "https://something.storage.host/upload/path/filename"
-	// }
-}
 
 func TestImageBlock_Validate(t *testing.T) {
 	v := validator.New()
